Rename useMessageNavigation hook file
